Refactor operation aggregation: replace running_ops_aggregation with operation_provider and remove unused operation_runtime_map
